Load the correct dictionary based on requests.
- Load correct dictionary based on server
  requests for HTML content with contractFile.
